move msal-service functions to authentication module file
